descarga de productos

+++ router/routerUsuarios.js
@@ -22,13 +22,16 @@ router.post("/listado", async(req,res)=>{
 // ELIMINA UN USUARIOS
 router.post("/eliminar", async(req,res)=>{
     
-    const {sucursal,id, codven} = req.body;
+    const {sucursal,id, codven,tipo} = req.body;
         
     let qry =''; let qryVendedor = '';
 
     qry = `DELETE FROM ME_USUARIOS WHERE CODSUCURSAL='${sucursal}'AND ID=${id};`;     
-    qryVendedor = `DELETE FROM ME_VENDEDORES WHERE CODSUCURSAL='${sucursal}' AND CODVEN=${codven};`
-
+    
+    if(tipo=='VENDEDOR'){
+        qryVendedor = `DELETE FROM ME_VENDEDORES WHERE CODSUCURSAL='${sucursal}' AND CODVEN=${codven};`
+    };
+    
     execute.Query(res, qry + qryVendedor);
 
 });
@@ -44,11 +47,14 @@ router.post("/nuevo", async(req,res)=>{
     qry = `INSERT INTO ME_USUARIOS 
         (CODUSUARIO,NOMBRE,PASS,TIPO,TELEFONO,CODDOC,CODSUCURSAL) VALUES
         (${codusuario},'${usuario}','${clave}','${tipo}','${telefono}','${coddoc}','${sucursal}');`;     
-        
+    
+    if(tipo=='VENDEDOR'){
+        qryV = `INSERT INTO ME_VENDEDORES 
+        (EMP_NIT,CODVEN,NOMVEN,ACTIVO,CODDOC,CODSUCURSAL) VALUES
+        ('1034261-3',${codusuario},'${usuario}','SI','${coddoc}','${sucursal}');`;    
+    };
 
-    qryV = `INSERT INTO ME_VENDEDORES 
-            (EMP_NIT,CODVEN,NOMVEN,ACTIVO,CODDOC,CODSUCURSAL) VALUES
-            ('1034261-3',${codusuario},'${usuario}','SI','${coddoc}','${sucursal}');`;    
+    
 
     execute.Query(res,qry + qryV);
 
@@ -70,7 +76,10 @@ router.post("/editar", async(req,res)=>{
             CODDOC='${coddoc}'
             WHERE ID=${id} AND CODSUCURSAL='${sucursal}';`;     
 
-    qryV = `UPDATE ME_VENDEDORES SET NOMVEN='${usuario}', CODDOC='${coddoc}' WHERE CODSUCURSAL='${sucursal}' AND CODVEN=${codusuario};`
+            if(tipo=='VENDEDOR'){
+                qryV = `UPDATE ME_VENDEDORES SET NOMVEN='${usuario}', CODDOC='${coddoc}' WHERE CODSUCURSAL='${sucursal}' AND CODVEN=${codusuario};`
+            }
+    
   
     execute.Query(res,qry + qryV);
